Reorganize FlureeQL import structure
