+++ src/index.tsx
@@ -1,5 +1,5 @@
 import { Editor } from './Components/Editor'
-import FlureeQL from './Views/FlureeQL/FlureeQL'
+import FlureeQL from './Views/FlureeQL'
 import { History } from './Components/History'
 import { flureeFetch } from './utils/flureeFetch'
